requirements.txt
tensorboard==2.1.0
tensorflow-estimator==2.1.0
tensorflow-gpu==2.1.0